requirements-dev.txt
# 开发环境依赖

# 代码质量工具
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.0.0

# 测试工具
pytest>=7.0.0
pytest-django>=4.5.0
pytest-cov>=4.0.0
coverage>=7.0.0
factory-boy>=3.2.0

# 调试工具
django-debug-toolbar>=4.0.0
django-extensions>=3.2.0
ipython>=8.0.0

# 文档工具
mkdocs>=1.4.0
mkdocs-material>=9.0.0
mkdocs-mermaid2-plugin>=0.6.0

# 代码检查
pre-commit>=3.0.0
bandit>=1.7.0
safety>=2.3.0

# API测试
httpie>=3.2.0
requests-mock>=1.10.0

# 性能分析
django-silk>=5.0.0
memory-profiler>=0.60.0

# 数据库工具
django-migration-linter>=4.1.0
